[#11] refactor: web - service - domain 계층적으로 패키지 구조 변경
- web - service - domain 계층적으로 패키지 구조 변경
