+++ src/main/java/dev/grabros/osikdang/domain/entity/myList/MyList.java
@@ -1,8 +1,9 @@
-package dev.grabros.osikdang.domain.favorite;
+package dev.grabros.osikdang.domain.entity.myList;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import dev.grabros.osikdang.domain.restaurant.Restaurant;
-import dev.grabros.osikdang.domain.user.User;
+import dev.grabros.osikdang.domain.entity.restaurant.Restaurant;
+import dev.grabros.osikdang.domain.entity.user.User;
+import java.time.LocalDate;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -18,11 +19,15 @@ import lombok.ToString;
 @Getter
 @ToString(exclude = {"user", "restaurant"})
 @NoArgsConstructor
-public class Favorite {
+public class MyList {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    private String feed;
+
+    private LocalDate modifiedTime;
 
     @JsonIgnore
     @ManyToOne(fetch = FetchType.LAZY)
