+++ src/main/java/dev/grabros/osikdang/domain/entity/restaurant/Restaurant.java
@@ -1,8 +1,8 @@
-package dev.grabros.osikdang.domain.restaurant;
+package dev.grabros.osikdang.domain.entity.restaurant;
 
-import dev.grabros.osikdang.domain.favorite.Favorite;
-import dev.grabros.osikdang.domain.myList.MyList;
-import dev.grabros.osikdang.domain.review.Review;
+import dev.grabros.osikdang.domain.entity.favorite.Favorite;
+import dev.grabros.osikdang.domain.entity.myList.MyList;
+import dev.grabros.osikdang.domain.entity.review.Review;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.CascadeType;
