+++ src/main/java/dev/grabros/osikdang/domain/entity/user/User.java
@@ -1,8 +1,8 @@
-package dev.grabros.osikdang.domain.user;
+package dev.grabros.osikdang.domain.entity.user;
 
-import dev.grabros.osikdang.domain.favorite.Favorite;
-import dev.grabros.osikdang.domain.myList.MyList;
-import dev.grabros.osikdang.domain.review.Review;
+import dev.grabros.osikdang.domain.entity.favorite.Favorite;
+import dev.grabros.osikdang.domain.entity.myList.MyList;
+import dev.grabros.osikdang.domain.entity.review.Review;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.CascadeType;
