+++ src/main/java/dev/grabros/osikdang/global/util/geometry/Location.java
@@ -1,4 +1,4 @@
-package dev.grabros.osikdang.geometry;
+package dev.grabros.osikdang.global.util.geometry;
 
 import lombok.Getter;
